relocate dice command
